Add modules for creating about, menu and contact content
