+++ src/about.js
@@ -1,4 +1,4 @@
-const about = () => {
+export default function about() {
   const item = document.querySelector('.item');
 
   const aboutTitle = document.createElement('h2');
@@ -12,25 +12,3 @@ const about = () => {
   item.appendChild(aboutTitle);
   item.appendChild(aboutText);
 };
-
-const menu = () => {
-
-};
-
-const contact = () => {
-  const item = document.querySelector('.item');
-
-  const aboutTitle = document.createElement('h2');
-  aboutTitle.classList.add('center');
-  aboutTitle.textContent = 'Hello';
-
-  const aboutText = document.createElement('p');
-  aboutText.classList.add('center');
-  aboutText.textContent = 'Hello';
-  
-  item.appendChild(aboutTitle);
-  item.appendChild(aboutText);
-};
-
-export default about;
-export {about, menu, contact};
