+++ src/index.js
@@ -1,5 +1,10 @@
 import * as home from './home'
+import about from './about'
+import menu from './menu'
+import contact from './contact'
 
 home.header();
 home.main();
 home.footer();
+
+about();
